# backend/openapi.yaml
openapi: 3.1.0
info:
  title: IoTSense API
  version: 1.0.0
  summary: Versioned facility monitoring and operations API.
  description: >-
    Initial contract. Authentication and health foundations are implemented; remaining
    business APIs are planned. Device credentials determine
    organization, washroom association and source mode. The frontend must not be treated
    as an authorization boundary. MQ readings are relative/raw values, not ppm.
servers:
  - url: https://api.example.invalid
    description: Placeholder production API origin
  - url: http://localhost:8000
    description: Local development API
...
paths:
  /api/v1/auth/login:
    post:
      tags: [Authentication]
      summary: Start a user session
      description: Sets a secure HttpOnly same-site session cookie. Apply throttling.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200':
          description: Authenticated user summary
          headers:
            Set-Cookie:
              schema: { type: string }
              description: Secure session cookie; never expose its value in JSON or logs.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserSummary' }
        '401': { $ref: '#/components/responses/Problem' }
        '429': { $ref: '#/components/responses/Problem' }
  /api/v1/auth/logout:
    post:
      tags: [Authentication]
      summary: Revoke current session
      security: [{ userSession: [] }]
      responses:
        '204': { description: Session revoked }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
  /api/v1/auth/me:
    get:
      tags: [Authentication]
      summary: Get current user and authorized scope
      security: [{ userSession: [] }]
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserSummary' }
        '401': { $ref: '#/components/responses/Problem' }
  /api/v1/auth/refresh:
    post:
      tags: [Authentication]
      summary: Rotate the active user session
      security: [{ userSession: [] }]
      parameters: [{ $ref: '#/components/parameters/CsrfToken' }]
      responses:
        '200':
          description: Rotated session and user summary
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserSummary' }
        '401': { $ref: '#/components/responses/Problem' }
  /api/v1/auth/password-reset/request:
    post:
      tags: [Authentication]
      summary: Request a password reset
      description: Always returns the same response to avoid account enumeration.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
      responses:
        '202': { description: Request accepted whether or not the account exists }
        '429': { $ref: '#/components/responses/Problem' }
  /api/v1/auth/password-reset/confirm:
    post:
      tags: [Authentication]
      summary: Set a new password using a one-time reset token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, new_password]
              properties:
                token: { type: string, writeOnly: true }
                new_password: { type: string, minLength: 12, maxLength: 256, writeOnly: true }
      responses:
        '204': { description: Password changed and previous sessions revoked }
        '400': { $ref: '#/components/responses/Problem' }
...
  /api/v1/users:
    get:
      tags: [Administration]
      summary: List users in authorized scope
      security: [{ userSession: [] }]
      parameters:
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/PageSize' }
      responses:
        '200':
          description: User page
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: '#/components/schemas/Page' }
                  - type: object
                    properties:
                      items:
                        type: array
                        items: { $ref: '#/components/schemas/UserSummary' }
        '403': { $ref: '#/components/responses/Problem' }
    post:
      tags: [Administration]
      summary: Invite a user into the permitted organization scope
      security: [{ userSession: [] }]
      parameters: [{ $ref: '#/components/parameters/CsrfToken' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, display_name, role]
              properties:
                email: { type: string, format: email }
                display_name: { type: string, minLength: 1, maxLength: 120 }
                role: { $ref: '#/components/schemas/Role' }
                organization_id: { type: [string, 'null'], format: uuid }
      responses:
        '201':
          description: Invitation created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserSummary' }
        '403': { $ref: '#/components/responses/Problem' }
  /api/v1/users/{userId}:
    delete:
      tags: [Administration]
      summary: Deactivate a user and revoke their sessions
      security: [{ userSession: [] }]
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - { $ref: '#/components/parameters/CsrfToken' }
      responses:
        '204': { description: User deactivated }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
...
components:
  securitySchemes:
    userSession:
      type: apiKey
      in: cookie
      name: iotsense_session
      description: Secure HttpOnly session cookie; browser mutations also require X-CSRF-Token.
    deviceKey:
      type: apiKey
      in: header
      name: X-Device-Key
      description: Unique revocable per-device key; server stores only a hash.
  parameters:
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: page_size
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
...
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema: { type: string }
  responses:
    Problem:
      description: RFC 9457-style API problem
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
    LoginRequest:
      type: object
      required: [email, password]
      additionalProperties: false
      properties:
        email: { type: string, format: email, maxLength: 254 }
        password: { type: string, minLength: 1, maxLength: 256, writeOnly: true }
    Role:
      type: string
      enum: [super_admin, facility_admin, maintenance_staff, viewer]
    UserSummary:
      type: object
      required: [id, email, display_name, role, status]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        display_name: { type: string }
        role: { $ref: '#/components/schemas/Role' }
        status: { type: string, enum: [invited, active, inactive] }
        organization_id: { type: [string, 'null'], format: uuid }
        facility_ids:
          type: array
          items: { type: string, format: uuid }
...
    Page:
      type: object
      required: [page, page_size, has_more]
      properties:
        page: { type: integer, minimum: 1 }
        page_size: { type: integer, minimum: 1, maximum: 200 }
        total: { type: integer, minimum: 0 }
        has_more: { type: boolean }
...
    Problem:
      type: object
      required: [type, title, status, code, request_id]
      properties:
        type: { type: string, format: uri-reference }
        title: { type: string }
        status: { type: integer, minimum: 400, maximum: 599 }
        code: { type: string }
        detail: { type: string }
        instance: { type: string }
        request_id: { type: string }
        errors:
          type: array
          items:
            type: object
            required: [field, message]
            properties:
              field: { type: string }
              message: { type: string }